+++ src/main/java/redlib/backend/vo/ScoreVO.java
@@ -7,12 +7,12 @@ import java.util.Date;
 @Data
 public class ScoreVO {
     private Integer id;
-    private Long studentNum;
-    private Byte chineseScore;
-    private Byte mathScore;
-    private Byte englishScore;
+    private Integer studentNum;
+    private Integer chineseScore;
+    private Integer mathScore;
+    private Integer englishScore;
     private Date entryEvent;
-    private Short academicYear;
-    private Byte semester;
+    private Integer academicYear;
+    private Integer semester;
     private Integer classId;
 }
